// backend/src/generated/prisma/enums.ts
export const OpWorkJobStatus = {
  DRAFT: 'DRAFT',
  ACTIVE: 'ACTIVE',
  PAUSED: 'PAUSED',
  CLOSED: 'CLOSED',
  ARCHIVED: 'ARCHIVED'
} as const

export type OpWorkJobStatus = (typeof OpWorkJobStatus)[keyof typeof OpWorkJobStatus]


export const OpWorkEmploymentType = {
  FULL_TIME: 'FULL_TIME',
  PART_TIME: 'PART_TIME',
  CONTRACT: 'CONTRACT',
  INTERNSHIP: 'INTERNSHIP',
  REMOTE: 'REMOTE',
  FREELANCE: 'FREELANCE'
} as const

export type OpWorkEmploymentType = (typeof OpWorkEmploymentType)[keyof typeof OpWorkEmploymentType]


export const OpWorkExperienceLevel = {
  INTERNSHIP: 'INTERNSHIP',
  ENTRY_LEVEL: 'ENTRY_LEVEL',
  JUNIOR: 'JUNIOR',
  MIDDLE: 'MIDDLE',
  SENIOR: 'SENIOR',
  LEAD: 'LEAD',
  EXPERT: 'EXPERT'
} as const

export type OpWorkExperienceLevel = (typeof OpWorkExperienceLevel)[keyof typeof OpWorkExperienceLevel]